+++ Changelog.md
@@ -3,6 +3,7 @@ OpenCore Changelog
 #### v0.6.3
 - Added support for xml comments in plist files
 - Updated underlying EDK II package to edk2-stable202008
+- Provide fallbacks for NULL memory SMBIOS strings
 
 #### v0.6.2
 - Updated builtin firmware versions for SMBIOS and the rest